app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "The avatar resizing service for Solus Orbis."
authors = ["Jessica Stamos <jessica@jessicastamos.com>"]

[tool.poetry.dependencies]
python = "^3.9.5"
boto3 = "^1.17.93"
fastapi = "^0.65.2"
Pillow = "^8.2.0"
python-multipart = "*"
requests = "^2.25.1"
uvicorn = "^0.14.0"

[tool.poetry.dev-dependencies]
black = "*"
python-dotenv = "*"